% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_by_column}
\alias{filter_by_column}
\title{Filter a data frame by column}
\usage{
filter_by_column(data, cols = NULL)
}
\arguments{
\item{data}{a data frame or similar object accepted by dplyr::filter}

\item{cols}{a vector of strings}
}
\value{
a data frame
}
\description{
Wrapper for dplyr::select that takes vector of a column names as strings and
returns a data frame containing only those columns. Retains the same column
ordering as the original data frame.
}
\examples{
filter_by_column(iris, cols = c("Sepal.Length", "Sepal.Width", "Species"))
}
