% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_by_row}
\alias{filter_by_row}
\title{Filter a data frame by row}
\usage{
filter_by_row(data, column_name, id = NULL)
}
\arguments{
\item{data}{a data frame or similar object accepted by dplyr::filter}

\item{column_name}{a string}

\item{id}{NULL or a string}
}
\value{
a data frame
}
\description{
Wrapper for dplyr::filter that takes a column name as a string and filters
for rows where that column matches the given `id`. If `id` is NULL or an
empty string then performs no filtering.
}
\examples{
filter_by_row(iris, column_name = "Species", id = "setosa")
}
