% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cols_to_format}
\alias{get_cols_to_format}
\title{Find columns matching a pattern}
\usage{
get_cols_to_format(data, cols = NULL)
}
\arguments{
\item{data}{a data frame}

\item{cols}{either a vector or column names or a regular expression
specifying column names to match. Will be treated as a regex if length is
1, otherwise as a vector of column names.}
}
\value{
a vector of column names
}
\description{
Find columns matching a pattern
}
\examples{
get_cols_to_format(iris, cols = "Length|Width")
}
