% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTable.R
\name{tableUI}
\alias{tableUI}
\title{Creates UI for a BioTable}
\usage{
tableUI(id, description_dir = "desc", helper = TRUE)
}
\arguments{
\item{id}{id for the module}

\item{description_dir}{optional, path to directory containing .md file with a
table description}

\item{helper}{logical, whether to add a helper element using the shinyhelper
package. If TRUE, looks for a file named `{id}.md` in a `helpfiles/`
directory}
}
\value{
a shiny::tagList
}
\description{
UI for a BioTable, which includes a table description, column selector, and
the table itself. Selecting columns in the column selector automatically
updates the columns that are included in the table.
}
