% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTable.R
\name{tableServer}
\alias{tableServer}
\title{Creates server for a BioTable}
\usage{
tableServer(
  id,
  data,
  row_id = shiny::reactive(NULL),
  id_column_name = NULL,
  default_cols = NULL
)
}
\arguments{
\item{id}{id for the module}

\item{data}{a data frame or similar object accepted by dplyr::filter/dplyr::select}

\item{row_id}{optional reactive element to use to filter rows}

\item{id_column_name}{optional column name to use for filtering}

\item{default_cols}{defaults columns to show in output}
}
\description{
Server function for a BioTable, which filters the table according to columns
selected in the UI (tableUI) and optionally according to a row filter which
can be shared across elements.
}
