% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTable.R
\name{tableUI}
\alias{tableUI}
\title{Creates UI for a BioTable}
\usage{
tableUI(id, md_description = TRUE, description_dir = "desc", helper = TRUE)
}
\arguments{
\item{id}{id for the module}

\item{md_description}{logical, whether to include a table description from a
markdown file. The file is expected to be named `desc/{id}.md`. The
directory can be specified using `description_dir`}

\item{description_dir}{optional, path to directory containing the table
description file}

\item{helper}{logical, whether to add a helper element using the shinyhelper
package. If TRUE, looks for a file named `{id}.md` in a `helpfiles/`
directory}
}
\value{
a shiny::tagList
}
\description{
UI for a BioTable, which includes a table description, column selector, and
the table itself. Selecting columns in the column selector automatically
updates the columns that are included in the table.
}
