% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTable.R
\name{tableUI}
\alias{tableUI}
\title{Creates UI for a BioTable}
\usage{
tableUI(id)
}
\arguments{
\item{id}{id for the module}
}
\value{
a shiny::tagList
}
\description{
UI for a BioTable, which includes a table description, column selector, and
the table itself. Selecting columns in the column selector automatically
updates the columns that are included in the table.
}
