% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTable.R
\name{tableUI}
\alias{tableUI}
\title{Creates UI for a BioTable}
\usage{
tableUI(id, all_cols, default_cols = NULL)
}
\arguments{
\item{id}{id for the module}

\item{all_cols}{character vector of all columns to list in the selection}

\item{default_cols}{columns to make visible by default}
}
\value{
a shiny::tagList
}
\description{
UI for a BioTable, which includes a table description, column selector, and
the table itself. Selecting columns in the column selector automatically
updates the columns that are included in the table.
}
